rename: 타입을 도메인 기준으로 재분류 feature/TODAK-197/update-mockup
