+++ app/(auth)/layout.tsx
@@ -2,7 +2,7 @@
 
 import { Logo } from "@/components/layout";
 import store from "@/redux";
-import { useMocking } from "@/utils/hooks";
+import { useMocking } from "@/domain/shared/hooks";
 import { HR } from "flowbite-react";
 import { Provider } from "react-redux";
