+++ components/my/UploadFileLabel.tsx
@@ -2,7 +2,7 @@
 
 import { RootState } from "@/redux";
 import { setSelectedFile } from "@/redux/slices/member/memberSlice";
-import { encodeFileToBase64 } from "@/utils/function";
+import { encodeFileToBase64 } from "@/domain/shared/function";
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
