+++ domain/diary/types/diaryRequestType.ts
