+++ domain/feed/types/feedResponseType.ts
