+++ domain/member/types/memberRequestType.ts
