+++ domain/shared/constants/index.ts
