+++ domain/shared/constants/initialConstants.ts
