+++ domain/shared/function/encodeBase64.ts
