+++ domain/shared/function/index.ts
