+++ domain/shared/hooks/useMocking.tsx
@@ -6,7 +6,7 @@ const useMocking = () => {
   useEffect(() => {
     if (process.env.NEXT_PUBLIC_API_MOCKING === "enabled") {
       (async () => {
-        const { initMocks } = await import("@/feature/mocks");
+        const { initMocks } = await import("@/mocks");
         await initMocks(); // initMocks를 비동기로 호출하여 초기화 완료
       })();
     }
