+++ domain/shared/lib/react-calendar/Calendar.css
