+++ domain/shared/routes/index.ts
