+++ domain/shared/types/common/animationType.ts
