+++ domain/shared/types/common/index.ts
